Brick break. Mario and Koopas can make brick break

+++ MarioGame/QuestionBrick.h
@@ -21,7 +21,7 @@
 class QuestionBrick : public CGameObject {
 
 	float ay;
-	int y_initial;
+	float y_initial;
 	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects);
 
 public:
